Add diff viewer
