+++ package.json
@@ -57,6 +57,7 @@
     "p-queue": "^7.3.4",
     "p-retry": "^5.1.2",
     "react": "^18.2.0",
+    "react-diff-viewer": "^3.1.1",
     "react-dom": "^18.2.0",
     "react-markdown": "^8.0.7",
     "snarkdown": "^2.0.0",
